+++ internal/entity/dtos/sql/supplyOrder/supply_order_item.go
@@ -15,13 +15,13 @@ type CreateIngredient struct {
 }
 
 type SupplyOrderItem struct {
-	SupplyOrderItemID uint    `gorm:"column:supplyorderitem_id;primaryKey;autoIncrement" json:"supplyorderitem_id"`
+	SupplyOrderItemID uint64  `gorm:"column:supplyorderitem_id;primaryKey;autoIncrement" json:"supplyorderitem_id"`
 	Price             float64 `gorm:"column:price;type:decimal(10,2);not null" json:"price"`
 	Quantity          int     `gorm:"column:quantity;not null" json:"quantity"`
 	Unit              string  `gorm:"column:unit;type:varchar(200);not null" json:"unit"`
-	IngredientID      uint    `gorm:"column:ingredient_id;not null;index" json:"ingredient_id"`
-	SupplyOrderID     uint    `gorm:"column:supplyorder_id;unique;not null;index" json:"supplyorder_id"`
-	StockBatchID      uint    `gorm:"column:stockbatch_id;not null;index" json:"stockbatch_id"`
+	IngredientID      uint64  `gorm:"column:ingredient_id;not null;index" json:"ingredient_id"`
+	SupplyOrderID     uint64  `gorm:"column:supplyorder_id;unique;not null;index" json:"supplyorder_id"`
+	StockBatchID      uint64  `gorm:"column:stockbatch_id;not null;index" json:"stockbatch_id"`
 
 	// Quan hệ với các bảng khác
 	//SupplyOrder SupplyOrder                `gorm:"belongsTo:SupplyOrder;foreignKey:SupplyOrderID;references:SupplyOrderID" json:"supply_order"`
